app/requirements.txt
annotated-types==0.6.0
azure-core==1.29.4
azure-identity==1.14.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
idna==3.4
isodate==0.6.1
msal==1.24.1
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
portalocker==2.8.2
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydo==0.1.7
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.6
